refactor: remaining branding issues
