+++ packages/frontend/index.html
@@ -19,11 +19,6 @@
     <script>
       window.global = window;
     </script>
-    <script
-      src="https://essential-eleven.sideshift.fi/script.js"
-      data-site="ATQUUJSG"
-      defer
-    ></script>
   </head>
   <body>
     <div id="root"></div>
