+++ packages/frontend/src/components/RecentWarps/RecentWarps.tsx
@@ -24,7 +24,7 @@ const RecentWarpsTableData: FunctionComponent = () => {
     return (
       <>
         {new Array(5).fill(0).map((_, index) => (
-          <Table.Row key={`recent-shifts-table-skeleton-${index}`}>
+          <Table.Row key={`recent-swaps-table-skeleton-${index}`}>
             <Table.Cell width="100%" className="text-center">
               <Skeleton className="sm:h-10 sm:w-100 m-auto rounded-md h-20" />
             </Table.Cell>
